requirements.txt
streamlit>=1.36.0
pandas>=2.2.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
openai>=1.40.0
pytest>=8.2.0
